refactor: rename onboarding components and update package structure for better organization
