+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/onBoarding/component/OnboardingCard.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.presentation.screens.onBoarding.onboarding
+package com.amsterdam.cutetudee.presentation.screens.onBoarding.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
